Refactor API routes and components for improved type safety and organization
- Updated type annotations in GET requests for user and order routes to use Promise for params.
- Changed import paths for BestSellingProduct and RecentOrder in components to use centralized lib.
- Added type definitions for default filters in FilterSidebar component.
- Enhanced state management in InventoryQuery component with explicit types.
- Improved performance metrics calculations in PerformanceMonitor component.
- Removed unnecessary props in SizeGuideModal within ProductInformation component.
- Updated RecentOrders component to use centralized lib for fetching recent orders.
- Implemented new login page with error handling and loading states.
- Created password reset page with token validation and user feedback.
- Developed request reset page for initiating password reset process.
- Added signup page with form validation and user feedback.
- Introduced dashboard lib for fetching and formatting dashboard data.
