+++ app/(user)/(auth)/request-reset/page.tsx
@@ -1,12 +1,12 @@
 
 "use client";
 
-import React, { useState, useEffect } from "react";
+import React, { useState, useEffect, Suspense } from "react";
 import Link from "next/link";
 import { useSearchParams } from "next/navigation";
 import { authNotifications } from "@/lib/notificationService";
 
-const RequestResetPage = () => {
+const RequestResetContent = () => {
   const [email, setEmail] = useState("");
   const [isLoading, setIsLoading] = useState(false);
   const [isSubmitted, setIsSubmitted] = useState(false);
@@ -113,6 +113,21 @@ const RequestResetPage = () => {
         )}
       </div>
     </div>
+  );
+};
+
+const RequestResetPage = () => {
+  return (
+    <Suspense fallback={
+      <div className="min-h-screen flex items-center justify-center bg-white">
+        <div className="text-center">
+          <div className="inline-block h-8 w-8 animate-spin rounded-full border-4 border-solid border-gray-300 border-r-transparent"></div>
+          <p className="mt-2 text-gray-600">Loading...</p>
+        </div>
+      </div>
+    }>
+      <RequestResetContent />
+    </Suspense>
   );
 };
