+++ app/admin/discountedit/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useState, useEffect } from "react";
+import { useState, useEffect, Suspense } from "react";
 import { BellIcon, Cog6ToothIcon, ClockIcon, MagnifyingGlassIcon } from "@heroicons/react/24/solid";
 import Sidebar from "../../components/Sidebar";
 import { useRouter, useSearchParams } from "next/navigation";
@@ -19,7 +19,7 @@ interface Category {
   thumbnailImage?: string;
 }
 
-export default function DiscountEdit() {
+function DiscountEditContent() {
   const router = useRouter();
   const searchParams = useSearchParams();
   const discountId = searchParams?.get("id");
@@ -557,5 +557,22 @@ export default function DiscountEdit() {
         </div>
       </div>
     </div>
+  );
+}
+
+export default function DiscountEdit() {
+  return (
+    <Suspense fallback={
+      <div className="flex min-h-screen bg-gray-100">
+        <div className="flex-1 flex justify-center items-center">
+          <div className="text-center">
+            <div className="inline-block h-8 w-8 animate-spin rounded-full border-4 border-solid border-orange-500 border-r-transparent"></div>
+            <p className="mt-2">Loading...</p>
+          </div>
+        </div>
+      </div>
+    }>
+      <DiscountEditContent />
+    </Suspense>
   );
 }
