+++ app/api/db-status/route.tsx
@@ -1,21 +1,23 @@
 import { NextResponse } from "next/server";
-import checkConnection from "@/lib/mongodb";
+import mongoose from "mongoose";
 
 export async function GET() {
   try {
-    const client = await checkConnection;
-    
-    if (client) {
+    // Check if we're already connected
+    if (mongoose.connection.readyState === 1) {
       return NextResponse.json({ 
         status: "connected", 
         message: "Database connection successful" 
       });
-    } else {
-      return NextResponse.json({ 
-        status: "disconnected", 
-        message: "Database connection failed" 
-      }, { status: 500 });
     }
+    
+    // Try to connect
+    await mongoose.connect(process.env.MONGODB_URI!);
+    
+    return NextResponse.json({ 
+      status: "connected", 
+      message: "Database connection successful" 
+    });
   } catch (error) {
     console.error("Error checking database connection:", error);
     return NextResponse.json({
