+++ app/api/discounts/[id]/route.tsx
@@ -18,7 +18,7 @@ const connectDB = async () => {
 // GET a discount by ID
 export async function GET(
   request: NextRequest,
-  { params }: { params: { id: string } }
+  { params }: { params: Promise<{ id: string }> }
 ) {
   try {
     await connectDB();
@@ -44,7 +44,7 @@ export async function GET(
 // PUT (update) a discount by ID
 export async function PUT(
   request: NextRequest,
-  { params }: { params: { id: string } }
+  { params }: { params: Promise<{ id: string }> }
 ) {
   try {
     await connectDB();
@@ -87,7 +87,7 @@ export async function PUT(
 // DELETE a discount by ID
 export async function DELETE(
   request: NextRequest,
-  { params }: { params: { id: string } }
+  { params }: { params: Promise<{ id: string }> }
 ) {
   try {
     await connectDB();
