+++ app/api/products/[id]/detail/route.tsx
@@ -20,7 +20,7 @@ const connectDB = async () => {
 // GET - Fetch complete product details with all related data
 export async function GET(
   request: NextRequest,
-  { params }: { params: { id: string } }
+  { params }: { params: Promise<{ id: string }> }
 ) {
   try {
     await connectDB();
