+++ app/api/products/[id]/route.tsx
@@ -39,7 +39,7 @@ const uploadToCloudinary = async (imageData: string) => {
 // GET a single product by ID
 export async function GET(
   request: NextRequest,
-  { params }: { params: { id: string } }
+  { params }: { params: Promise<{ id: string }> }
 ) {
   try {
     await connectDB();
@@ -73,7 +73,7 @@ export async function GET(
 // UPDATE a product by ID
 export async function PUT(
   request: NextRequest,
-  { params }: { params: { id: string } }
+  { params }: { params: Promise<{ id: string }> }
 ) {
   try {
     await connectDB();
@@ -203,7 +203,7 @@ export async function PUT(
 // DELETE a product by ID
 export async function DELETE(
   request: NextRequest,
-  { params }: { params: { id: string } }
+  { params }: { params: Promise<{ id: string }> }
 ) {
   try {
     await connectDB();
